# [PERF] hr_timesheet: speed up portal access to timesheets