@robodoo 2024-07-05
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/172130.png)](https://mergebot.odoo.com/odoo/odoo/pull/172130)